requirements.txt
NudeNet == 2.0.6
Pillow == 8.0.1
